Move raw2userData out to its own utility file
